+++ App.js
@@ -15,6 +15,8 @@ import Badge from './example/Badge';
 import Col from './example/Col';
 import Tag from './example/Tag';
 import Text from './example/Text';
+import Input from './example/Input';
+import Textarea from './example/Textarea';
 
 const AppNavigator = createStackNavigator({
   Example: {
@@ -43,6 +45,12 @@ const AppNavigator = createStackNavigator({
   },
   Text: {
     screen: Text,
+  },
+  Input: {
+    screen: Input,
+  },
+  Textarea: {
+    screen: Textarea,
   },
 }, {
   initialRouteName: 'Example',