+++ CHANGELOG.md
@@ -1,5 +1,27 @@
 # CHANGELOG
 
+## Version 0.0.5
+Build: 191201
+
+### Updated
+
+- Add Text component
+- Add Textarea component
+- Update logo project and app :tada:
+- Update layout Example
+
+
+
+## Version 0.0.4
+Build: 191130
+
+### Updated
+
+- Add Col component
+- Add Tag component
+- Add Text component
+
+
 ## Version 0.0.3
 Build: 191129